+++ lib/openapi_parser/spec_validator.rb
@@ -32,6 +32,7 @@ require_relative 'spec_validator/rules/additional_operations_before_32'
 require_relative 'spec_validator/rules/querystring_before_32'
 require_relative 'spec_validator/rules/cookie_style_before_32'
 require_relative 'spec_validator/rules/response_summary_before_32'
+require_relative 'spec_validator/rules/allow_empty_value_deprecation'
 
 module OpenAPIParser
   class SpecViolationError < OpenAPIError
@@ -111,6 +112,7 @@ module OpenAPIParser
           Rules::QuerystringBefore32,
           Rules::CookieStyleBefore32,
           Rules::ResponseSummaryBefore32,
+          Rules::AllowEmptyValueDeprecation,
         ]
       end
   end